style: removed rowGap as it is not supported in some browsers
